Move spark to its own module
